feat: 11-handler-refresh-token

+++ src/services/keyToken.service.js
@@ -29,14 +29,25 @@ class KeyTokenService {
   }
 
   static findByUserId = async (userId) => {
-    const objectId = new Types.ObjectId(userId)
+    console.log('userId: ', userId)
+    const objectId = new Types.ObjectId(userId) // !TODO: refactor this
     return await keyTokenModel.findOne({ user: objectId }).lean()
   }
 
   static removeKeyById = async (id) => {
-    return await keyTokenModel.deleteOne({
-      _id: id
-    })
+    return await keyTokenModel.deleteOne({ _id: id })
+  }
+
+  static findByRefreshTokenUsed = async (refreshToken) => {
+    return await keyTokenModel.findOne({ refreshTokensUsed: refreshToken })
+  }
+
+  static findByRefreshToken = async (refreshToken) => {
+    return await keyTokenModel.findOne({ refreshToken })
+  }
+
+  static deleteKeyById = async (userId) => {
+    return await keyTokenModel.deleteOne({ user: new Types.ObjectId(userId) })
   }
 }
